+++ packages-js/ui-core/src/identifiers/index.ts
@@ -1,0 +1,5 @@
+export { identifierMessages, type IdentifierMessageKey } from './messages'
+export * from './model'
+export { createIdentifiersRestPort } from './rest-port'
+export type * from './types'
+export * from './controller'